backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
weaviate-client>=3.26.7,<4.0.0
openai==1.12.0
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.2.1
python-magic==0.4.27
PyPDF2==3.0.1
docx2txt==0.8
tiktoken==0.9.0
